Fix incorrect types for casting pointers (was using long long instead of size_t so it broke on 32bit). Add save success indicator

+++ src/hooks/EditLevelLayer.cpp
@@ -4,15 +4,15 @@
 using namespace geode::prelude;
 using namespace persistenceAPI;
 
-unsigned long long s_editLevelLayerDelegate1 = 0;
-unsigned long long s_editLevelLayerDelegate2 = 0;
+size_t s_editLevelLayerDelegate1 = 0;
+size_t s_editLevelLayerDelegate2 = 0;
 
 // overrides
 
 void PSEditLevelLayer::onPlay(cocos2d::CCObject* i_sender) {
 	// stupid way of not letting it unregister the keyboard delegates
-	s_editLevelLayerDelegate1 = reinterpret_cast<unsigned long long>(&m_bTouchEnabled)-3*sizeof(void*);
-	s_editLevelLayerDelegate2 = reinterpret_cast<unsigned long long>(&m_bTouchEnabled)-2*sizeof(void*);
+	s_editLevelLayerDelegate1 = reinterpret_cast<size_t>(&m_bTouchEnabled)-3*sizeof(void*);
+	s_editLevelLayerDelegate2 = reinterpret_cast<size_t>(&m_bTouchEnabled)-2*sizeof(void*);
 
 	PSFMODAudioEngine* l_audioEngine = static_cast<PSFMODAudioEngine*>(FMODAudioEngine::get());
 	l_audioEngine->m_fields->m_disableLoadMusic = true;
