+++ src/hooks/EditLevelLayer.hpp
@@ -3,8 +3,8 @@
 #include <Geode/modify/EditLevelLayer.hpp>
 #include <sabe.persistenceapi/include/PersistenceAPI.hpp>
 
-extern unsigned long long s_editLevelLayerDelegate1;
-extern unsigned long long s_editLevelLayerDelegate2;
+extern size_t s_editLevelLayerDelegate1;
+extern size_t s_editLevelLayerDelegate2;
 
 class $modify(PSEditLevelLayer, EditLevelLayer) {
 public:
