+++ src/hooks/cocos2d/CCKeyboardDispatcher.cpp
@@ -8,10 +8,10 @@ using namespace persistenceAPI;
 // overrides
 
 void PSCCKeyboardDispatcher::removeDelegate(cocos2d::CCKeyboardDelegate* i_delegate) {
-	if (s_levelInfoLayerDelegate1 == reinterpret_cast<unsigned long long>(i_delegate)
-		|| s_levelInfoLayerDelegate2 == reinterpret_cast<unsigned long long>(i_delegate)
-		|| s_editLevelLayerDelegate1 == reinterpret_cast<unsigned long long>(i_delegate)
-		|| s_editLevelLayerDelegate2 == reinterpret_cast<unsigned long long>(i_delegate)
+	if (s_levelInfoLayerDelegate1 == reinterpret_cast<size_t>(i_delegate)
+		|| s_levelInfoLayerDelegate2 == reinterpret_cast<size_t>(i_delegate)
+		|| s_editLevelLayerDelegate1 == reinterpret_cast<size_t>(i_delegate)
+		|| s_editLevelLayerDelegate2 == reinterpret_cast<size_t>(i_delegate)
 	) {
 		return;
 	}
